#80 refactor common fixtures to conftest.py
